Extracted create dialog into a component
