+++ src/components/Settings/index.tsx
@@ -1,36 +1,24 @@
-import React, { useState } from "react";
+import React from "react";
 import { useTheme } from "@mui/material/styles";
 import Paper from "@mui/material/Paper";
 import Stack from "@mui/material/Stack";
 import Typography from "@mui/material/Typography";
-import Dialog from "@mui/material/Dialog";
-import TextField from "@mui/material/TextField";
-import Collapse from "@mui/material/Collapse";
 import EditOutlinedIcon from "@mui/icons-material/EditOutlined";
 import DeleteOutlineOutlinedIcon from "@mui/icons-material/DeleteOutlineOutlined";
 import AddCircleOutlineOutlinedIcon from "@mui/icons-material/AddCircleOutlineOutlined";
-import AddOutlinedIcon from "@mui/icons-material/AddOutlined";
 
 import { useAppSelector, useFloors } from "../../redux/hooks";
+
 import DrawerAdjacent from "../m3/DrawerAdjacent";
 import M3IconButton from "../m3/M3IconButton";
-import { SurfaceTint } from "../m3/Tints";
-import M3Fab from "../m3/M3Fab";
-import M3TextButton from "../m3/M3TextButton";
-import Box from "@mui/material/Box";
+import CreateFloorDialog from "./CreateFloorDialog";
 
 export default function Settings(): JSX.Element {
   const theme = useTheme();
   const floors = useFloors();
   const roomTypes = useAppSelector((state) => state.roomTypes.data);
-  const [creatingState, setCreatingState] = useState<"idle" | "creating" | "loading">("idle");
-
   const floorIds = Object.keys(floors);
-  const openDialog = creatingState !== "idle";
 
-  function closeDialog(): void {
-    setCreatingState("idle");
-  }
 
   return (
     <DrawerAdjacent>
@@ -94,62 +82,7 @@ export default function Settings(): JSX.Element {
           })}
         </Stack>
       </Stack>
-      <M3Fab onClick={() => setCreatingState("creating")} sx={{
-        position: "fixed",
-        right: 0,
-        bottom: 0,
-        width: "auto",
-        minWidth: "5rem"
-      }}>
-        <Stack direction="row" spacing={1.5} alignItems="center" sx={{ p: "1rem" }}>
-          <AddOutlinedIcon />
-          <Typography variant="labelLarge">Crea</Typography>
-        </Stack>
-        <SurfaceTint sx={{
-          backgroundColor: theme.palette.primary.light,
-          opacity: theme.opacities.surface3
-        }} />
-      </M3Fab>
-      <Dialog keepMounted open={openDialog} onClose={closeDialog} PaperProps={{
-        elevation: 0,
-        sx: {
-          height: "calc(100vh - 4rem)",
-          backgroundColor: "transparent",
-          overflow: "visible"
-        }
-      }}>
-        <Paper elevation={3} sx={{
-          borderRadius: "1.75rem",
-          minWidth: "17.5rem",
-          maxWidth: "35rem",
-          position: "relative",
-          mt: "calc(50vh - 11.125rem)"
-        }}>
-          <Collapse
-            in={openDialog}
-            easing={theme.transitions.easing.fastOutSlowIn}
-          >
-            <Stack spacing={3} sx={{ p: "1.5rem" }}>
-              <Stack spacing={2} alignItems="center">
-                <AddOutlinedIcon />
-                <Typography variant="headlineSmall">Crea piano</Typography>
-                <Typography variant="bodyMedium" sx={{ display: "block", width: "100%", textAlign: "left" }}>
-                  Crea un nuovo piano vuoto con seguente nome:
-                </Typography>
-                <TextField label="Nome" sx={{ minWidth: "20rem" }} />
-              </Stack>
-              <Stack direction="row" spacing={1} justifyContent="flex-end">
-                <M3TextButton onClick={closeDialog}>Cancella</M3TextButton>
-                <M3TextButton>Crea</M3TextButton>
-              </Stack>
-            </Stack>
-          </Collapse>
-          <SurfaceTint sx={{
-            backgroundColor: theme.palette.primary.light,
-            opacity: theme.opacities.surface3
-          }} />
-        </Paper>
-      </Dialog>
+      <CreateFloorDialog />
     </DrawerAdjacent>
   );
 }
